refactor billing services
