+++ jumpscale/packages/vdc/billing.py
@@ -1,4 +1,5 @@
 import os
+import requests
 from decimal import Decimal
 from jumpscale.loader import j
 from jumpscale.sals.vdc.size import VDC_FLAVORS
@@ -8,11 +9,20 @@ from jumpscale.sals.vdc.models import K8SNodeFlavor
 
 PROVISION_WALLET_NAME = "provisioning"
 PREPAID_WALLET = "prepaid"
-ORIGINAL_USER_PLAN_TFT = int(os.getenv("ORIGINAL_USER_PLAN_TFT", 0))
-BASE_CAPACITY = int(os.getenv("BASE_CAPACITY", 0))
+BASE_CAPACITY = int(os.getenv("BASE_CAPACITY", 14))
 
 
 def get_addons(flavor, kubernetes_addons):
+    """Get all the addons on the basic user plan
+
+    Args:
+        flavor(str): user flavor for the plan
+        kubernetes_addons(list): all kubernetes nodes 
+    
+    Returns:
+        addons(list): list of addons used by the user over the basic chosen plan
+
+    """
     plan = VDC_FLAVORS.get(flavor)
     plan_nodes_count = plan.get("k8s").get("no_nodes")
     plan_nodes_size = plan.get("k8s").get("size")
@@ -28,16 +38,40 @@ def get_addons(flavor, kubernetes_addons):
     return addons
 
 
+def get_prices():
+    """Get the prices of the addons and plans from github 
+    Returns:
+        VDC_PRICES(dict): return VDC_prices dictionary form out config
+    """
+    # Store prices from github
+    if not j.config.get("VDC_PRICES"):
+        prices = requests.get(
+            "https://raw.githubusercontent.com/threefoldfoundation/vdc_pricing/master/prices.json"
+        ).json()
+        j.config.set("VDC_PRICES", prices)
+    return j.config.get("VDC_PRICES")
+
+
 def calculate_addon_price(addon):
-    prices = j.config.get("VDC_PRICES")
+    """Calculate the price of a single addon
+    Args:
+        addon(obj): addon object 
+    Returns:
+        (float): addon price
+    """
+    prices = get_prices()
     size = addon.size.name.lower()
     return prices["nodes"][size]
 
 
 def calculate_plan_base_price():
+    """Get user plan price
+    Returns:
+        float(): the price of user plan
+    """
     vdc_instance_name = os.getenv("VDC_INSTANCE_NAME")
     vdc_instance = j.sals.vdc.get(vdc_instance_name)
-    prices = j.config.get("VDC_PRICES")
+    prices = get_prices()
     return prices["plans"][vdc_instance.flavor.value]
 
 
@@ -67,7 +101,8 @@ def calculate_hourly_rate():
         hourly_amount(float): the total price for each hour,
                               including the price of the user plan and addons
     """
-    hourly_amount = ORIGINAL_USER_PLAN_TFT / (24 * 30)
+    user_plan_price = calculate_plan_base_price()
+    hourly_amount = user_plan_price / (24 * 30)
     hourly_amount += calculate_addons_hourly_rate()
     return hourly_amount
 
@@ -86,7 +121,7 @@ def auto_extend_billing():
     """Is used to get the pool in the VDC and extend them when the remaining time is less than
     half of the BASE_CAPACITY
     """
-    # TODO: get the VDC pool
+    # Get the VDC and deployer instances
     vdc_instance_name = os.getenv("VDC_INSTANCE_NAME")
     vdc_password = os.getenv("VDC_PASSWORD")
     vdc_instance = j.sals.vdc.get(vdc_instance_name)
